+++ src/main/java/cn/renlm/graph/modular/er/entity/ErFieldLib.java
@@ -16,7 +16,7 @@ import lombok.experimental.Accessors;
  * </p>
  *
  * @author Renlm
- * @since 2022-05-16
+ * @since 2022-06-08
  */
 @Getter
 @Setter
@@ -66,7 +66,7 @@ public class ErFieldLib implements Serializable {
      * 长度
      */
     @TableField("size")
-    private Integer size;
+    private Long size;
 
     /**
      * 精度
